test: extract common logic and combine HRC904 solidity code
