+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/spec/dsl/contracts/TokenRedirectContract.java
@@ -22,10 +22,7 @@ package com.hedera.services.bdd.spec.dsl.contracts;
  */
 public enum TokenRedirectContract {
     HRC("HRC"),
-    // TODO: Update this to HRC904 once all tests are merged
-    HRC904CLAIM("HRC904TokenClaim"),
-    HRC904CANCEL("HRC904TokenCancel"),
-    HRC904REJECT("HRC904Reject"),
+    HRC904("HRC904"),
     ERC20("ERC20ABI"),
     ERC721("ERC721ABI");
