+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/precompile/airdrops/HRCTokenRejectTest.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.hedera.services.bdd.suites.contract.precompile;
+package com.hedera.services.bdd.suites.contract.precompile.airdrops;
 
 import static com.hedera.services.bdd.junit.TestTags.SMART_CONTRACT;
 import static com.hedera.services.bdd.spec.HapiSpec.hapiTest;
-import static com.hedera.services.bdd.spec.dsl.contracts.TokenRedirectContract.HRC904REJECT;
+import static com.hedera.services.bdd.spec.dsl.contracts.TokenRedirectContract.HRC904;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.CONTRACT_REVERT_EXECUTED;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INSUFFICIENT_TOKEN_BALANCE;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_OWNER_ID;
@@ -60,7 +60,7 @@ public class HRCTokenRejectTest {
                 token.treasury().transferUnitsTo(sender, 10L, token),
                 token.treasury().getBalance().andAssert(balance -> balance.hasTokenBalance(token.name(), 990L)),
                 sender.getBalance().andAssert(balance -> balance.hasTokenBalance(token.name(), 10L)),
-                token.call(HRC904REJECT, "rejectTokenFT").with(call -> call.payingWith(sender.name())),
+                token.call(HRC904, "rejectTokenFT").with(call -> call.payingWith(sender.name())),
                 sender.getBalance().andAssert(balance -> balance.hasTokenBalance(token.name(), 0L)),
                 token.treasury().getBalance().andAssert(balance -> balance.hasTokenBalance(token.name(), 1000L)));
     }
@@ -73,7 +73,7 @@ public class HRCTokenRejectTest {
                 nft.treasury().transferNFTsTo(sender, nft, 1L),
                 nft.treasury().getBalance().andAssert(balance -> balance.hasTokenBalance(nft.name(), 0L)),
                 sender.getBalance().andAssert(balance -> balance.hasTokenBalance(nft.name(), 1L)),
-                nft.call(HRC904REJECT, "rejectTokenNFTs", new long[] {1L}).with(call -> call.payingWith(sender.name())),
+                nft.call(HRC904, "rejectTokenNFTs", new long[] {1L}).with(call -> call.payingWith(sender.name())),
                 sender.getBalance().andAssert(balance -> balance.hasTokenBalance(nft.name(), 0L)),
                 nft.treasury().getBalance().andAssert(balance -> balance.hasTokenBalance(nft.name(), 1L)));
     }
@@ -87,7 +87,7 @@ public class HRCTokenRejectTest {
                 nft.treasury().transferNFTsTo(sender, nft, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L),
                 nft.treasury().getBalance().andAssert(balance -> balance.hasTokenBalance(nft.name(), 0L)),
                 sender.getBalance().andAssert(balance -> balance.hasTokenBalance(nft.name(), 10L)),
-                nft.call(HRC904REJECT, "rejectTokenNFTs", new long[] {1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L})
+                nft.call(HRC904, "rejectTokenNFTs", new long[] {1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L})
                         .with(call -> call.payingWith(sender.name()))
                         .gas(1_000_000L),
                 sender.getBalance().andAssert(balance -> balance.hasTokenBalance(nft.name(), 0L)),
@@ -101,7 +101,7 @@ public class HRCTokenRejectTest {
         return hapiTest(
                 sender.associateTokens(nft),
                 sender.getBalance().andAssert(balance -> balance.hasTokenBalance(nft.name(), 0L)),
-                nft.call(HRC904REJECT, "rejectTokenNFTs", new long[] {1L})
+                nft.call(HRC904, "rejectTokenNFTs", new long[] {1L})
                         .with(call -> call.payingWith(sender.name()))
                         .andAssert(txn -> txn.hasKnownStatuses(SUCCESS, INVALID_OWNER_ID)));
     }
@@ -112,7 +112,7 @@ public class HRCTokenRejectTest {
         return hapiTest(
                 sender.associateTokens(token),
                 sender.getBalance().andAssert(balance -> balance.hasTokenBalance(token.name(), 0L)),
-                token.call(HRC904REJECT, "rejectTokenFT")
+                token.call(HRC904, "rejectTokenFT")
                         .with(call -> call.payingWith(sender.name()))
                         .andAssert(txn -> txn.hasKnownStatuses(SUCCESS, INSUFFICIENT_TOKEN_BALANCE)));
     }
@@ -126,7 +126,7 @@ public class HRCTokenRejectTest {
                 nft.treasury().transferNFTsTo(sender, nft, 1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L),
                 nft.treasury().transferNFTsTo(sender, nft, 11L),
                 sender.getBalance().andAssert(balance -> balance.hasTokenBalance(nft.name(), 11L)),
-                nft.call(HRC904REJECT, "rejectTokenNFTs", new long[] {1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L, 11L})
+                nft.call(HRC904, "rejectTokenNFTs", new long[] {1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 9L, 10L, 11L})
                         .with(call -> call.payingWith(sender.name()))
                         .gas(1_000_000L)
                         .andAssert(txn -> txn.hasKnownStatuses(CONTRACT_REVERT_EXECUTED)));
