+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/precompile/airdrops/SystemContractAirdropHelper.java
@@ -1,0 +1,145 @@
+/*
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hedera.services.bdd.suites.contract.precompile.airdrops;
+
+import static com.hedera.services.bdd.spec.assertions.TransactionRecordAsserts.includingFungiblePendingAirdrop;
+import static com.hedera.services.bdd.spec.assertions.TransactionRecordAsserts.includingNftPendingAirdrop;
+import static com.hedera.services.bdd.spec.assertions.TransactionRecordAsserts.recordWith;
+import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTxnRecord;
+import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenAirdrop;
+import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.moving;
+import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.movingUnique;
+import static com.hedera.services.bdd.spec.utilops.CustomSpecAssert.allRunFor;
+
+import com.esaulpaugh.headlong.abi.Address;
+import com.hedera.services.bdd.spec.HapiSpec;
+import com.hedera.services.bdd.spec.SpecOperation;
+import com.hedera.services.bdd.spec.dsl.entities.SpecAccount;
+import com.hedera.services.bdd.spec.dsl.entities.SpecFungibleToken;
+import com.hedera.services.bdd.spec.dsl.entities.SpecNonFungibleToken;
+import com.hedera.services.bdd.spec.dsl.operations.queries.GetBalanceOperation;
+import com.hedera.services.bdd.spec.transactions.token.TokenMovement;
+import edu.umd.cs.findbugs.annotations.NonNull;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
+import java.util.stream.Stream;
+
+public final class SystemContractAirdropHelper {
+
+    public static GetBalanceOperation checkForEmptyBalance(
+            final SpecAccount receiver, final List<SpecFungibleToken> tokens, final List<SpecNonFungibleToken> nfts) {
+        return receiver.getBalance().andAssert(balance -> {
+            tokens.forEach(token -> balance.hasTokenBalance(token.name(), 0L));
+            nfts.forEach(nft -> balance.hasTokenBalance(nft.name(), 0L));
+        });
+    }
+
+    public static Address[] prepareSenderAddresses(@NonNull HapiSpec spec, @NonNull SpecAccount... senders) {
+        return Arrays.stream(senders)
+                .map(sender -> sender.addressOn(spec.targetNetworkOrThrow()))
+                .toArray(Address[]::new);
+    }
+
+    public static Address[] prepareReceiverAddresses(@NonNull HapiSpec spec, @NonNull SpecAccount... receivers) {
+        return Arrays.stream(receivers)
+                .map(receiver -> receiver.addressOn(spec.targetNetworkOrThrow()))
+                .toArray(Address[]::new);
+    }
+
+    public static Address[] prepareTokenAddresses(@NonNull HapiSpec spec, @NonNull SpecFungibleToken... tokens) {
+        return Arrays.stream(tokens)
+                .map(token -> token.addressOn(spec.targetNetworkOrThrow()))
+                .toArray(Address[]::new);
+    }
+
+    public static Address[] prepareNftAddresses(@NonNull HapiSpec spec, @NonNull SpecNonFungibleToken... nfts) {
+        return Arrays.stream(nfts)
+                .map(nft -> nft.addressOn(spec.targetNetworkOrThrow()))
+                .toArray(Address[]::new);
+    }
+
+    public static List<TokenMovement> prepareFTAirdrops(
+            @NonNull final SpecAccount sender,
+            @NonNull final SpecAccount receiver,
+            @NonNull final List<SpecFungibleToken> tokens) {
+        return tokens.stream()
+                .map(token -> moving(10, token.name()).between(sender.name(), receiver.name()))
+                .toList();
+    }
+
+    public static List<TokenMovement> prepareNFTAirdrops(
+            @NonNull final SpecAccount sender,
+            @NonNull final SpecAccount receiver,
+            @NonNull final List<SpecNonFungibleToken> nfts) {
+        return nfts.stream()
+                .map(nft -> movingUnique(nft.name(), 1L).between(sender.name(), receiver.name()))
+                .toList();
+    }
+
+    public static GetBalanceOperation checkForBalances(
+            final SpecAccount receiver, final List<SpecFungibleToken> tokens, final List<SpecNonFungibleToken> nfts) {
+        return receiver.getBalance().andAssert(balance -> {
+            tokens.forEach(token -> balance.hasTokenBalance(token.name(), 10L));
+            nfts.forEach(nft -> balance.hasTokenBalance(nft.name(), 1L));
+        });
+    }
+
+    public static void prepareAirdrops(
+            final SpecAccount sender,
+            final SpecAccount receiver,
+            @NonNull List<SpecFungibleToken> tokens,
+            @NonNull List<SpecNonFungibleToken> nfts,
+            @NonNull HapiSpec spec) {
+        var tokenMovements = prepareFTAirdrops(sender, receiver, tokens);
+        var nftMovements = prepareNFTAirdrops(sender, receiver, nfts);
+        allRunFor(
+                spec,
+                tokenAirdrop(Stream.of(tokenMovements, nftMovements)
+                                .flatMap(Collection::stream)
+                                .toArray(TokenMovement[]::new))
+                        .payingWith(sender.name())
+                        .via("tokenAirdrop"),
+                getTxnRecord("tokenAirdrop")
+                        .hasPriority(recordWith()
+                                .pendingAirdrops(
+                                        includingFungiblePendingAirdrop(tokenMovements.toArray(TokenMovement[]::new)))
+                                .pendingAirdrops(
+                                        includingNftPendingAirdrop(nftMovements.toArray(TokenMovement[]::new)))));
+    }
+
+    public static SpecOperation[] prepareTokensAndBalances(
+            final SpecAccount sender,
+            final SpecAccount receiver,
+            final List<SpecFungibleToken> tokens,
+            final List<SpecNonFungibleToken> nfts) {
+        ArrayList<SpecOperation> specOperations = new ArrayList<>();
+        specOperations.addAll(List.of(
+                sender.associateTokens(tokens.toArray(SpecFungibleToken[]::new)),
+                sender.associateTokens(nfts.toArray(SpecNonFungibleToken[]::new)),
+                checkForEmptyBalance(receiver, tokens, nfts)));
+        specOperations.addAll(tokens.stream()
+                .map(token -> token.treasury().transferUnitsTo(sender, 1_000L, token))
+                .toList());
+        specOperations.addAll(nfts.stream()
+                .map(nft -> nft.treasury().transferNFTsTo(sender, nft, 1L))
+                .toList());
+
+        return specOperations.toArray(SpecOperation[]::new);
+    }
+}
