+++ hedera-node/test-clients/src/main/resources/contract/contracts/HRC904/HRC904.bin
@@ -1,0 +1,1 @@
+608060405234801561001057600080fd5b50610a7a806100206000396000f3fe608060405234801561001057600080fd5b506004361061007d5760003560e01c8063a869c78a1161005b578063a869c78a14610100578063ad4917cf14610130578063cef5b70514610160578063f5677e99146101905761007d565b806363ada5d71461008257806376c6b391146100b2578063a83bc5b2146100d0575b600080fd5b61009c60048036038101906100979190610607565b6101c0565b6040516100a99190610656565b60405180910390f35b6100ba610248565b6040516100c79190610656565b60405180910390f35b6100ea60048036038101906100e59190610671565b6102c0565b6040516100f79190610656565b60405180910390f35b61011a600480360381019061011591906107f7565b610345565b6040516101279190610656565b60405180910390f35b61014a60048036038101906101459190610607565b6103ca565b6040516101579190610656565b60405180910390f35b61017a60048036038101906101759190610671565b610452565b6040516101879190610656565b60405180910390f35b6101aa60048036038101906101a59190610878565b6104d7565b6040516101b79190610656565b60405180910390f35b60003073ffffffffffffffffffffffffffffffffffffffff166363ada5d784846040518363ffffffff1660e01b81526004016101fd9291906108b4565b6020604051808303816000875af115801561021c573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061024091906108f2565b905092915050565b60003073ffffffffffffffffffffffffffffffffffffffff166376c6b3916040518163ffffffff1660e01b81526004016020604051808303816000875af1158015610297573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906102bb91906108f2565b905090565b60003073ffffffffffffffffffffffffffffffffffffffff1663a83bc5b2836040518263ffffffff1660e01b81526004016102fb919061091f565b6020604051808303816000875af115801561031a573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061033e91906108f2565b9050919050565b60003073ffffffffffffffffffffffffffffffffffffffff1663a869c78a836040518263ffffffff1660e01b815260040161038091906109f8565b6020604051808303816000875af115801561039f573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906103c391906108f2565b9050919050565b60003073ffffffffffffffffffffffffffffffffffffffff1663ad4917cf84846040518363ffffffff1660e01b81526004016104079291906108b4565b6020604051808303816000875af1158015610426573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061044a91906108f2565b905092915050565b60003073ffffffffffffffffffffffffffffffffffffffff1663cef5b705836040518263ffffffff1660e01b815260040161048d919061091f565b6020604051808303816000875af11580156104ac573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906104d091906108f2565b9050919050565b60003073ffffffffffffffffffffffffffffffffffffffff1663f5677e99836040518263ffffffff1660e01b81526004016105129190610a29565b6020604051808303816000875af1158015610531573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061055591906108f2565b9050919050565b6000604051905090565b600080fd5b600080fd5b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b600061059b82610570565b9050919050565b6105ab81610590565b81146105b657600080fd5b50565b6000813590506105c8816105a2565b92915050565b60008160070b9050919050565b6105e4816105ce565b81146105ef57600080fd5b50565b600081359050610601816105db565b92915050565b6000806040838503121561061e5761061d610566565b5b600061062c858286016105b9565b925050602061063d858286016105f2565b9150509250929050565b610650816105ce565b82525050565b600060208201905061066b6000830184610647565b92915050565b60006020828403121561068757610686610566565b5b6000610695848285016105b9565b91505092915050565b600080fd5b6000601f19601f8301169050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b6106ec826106a3565b810181811067ffffffffffffffff8211171561070b5761070a6106b4565b5b80604052505050565b600061071e61055c565b905061072a82826106e3565b919050565b600067ffffffffffffffff82111561074a576107496106b4565b5b602082029050602081019050919050565b600080fd5b600061077361076e8461072f565b610714565b905080838252602082019050602084028301858111156107965761079561075b565b5b835b818110156107bf57806107ab88826105f2565b845260208401935050602081019050610798565b5050509392505050565b600082601f8301126107de576107dd61069e565b5b81356107ee848260208601610760565b91505092915050565b60006020828403121561080d5761080c610566565b5b600082013567ffffffffffffffff81111561082b5761082a61056b565b5b610837848285016107c9565b91505092915050565b60008115159050919050565b61085581610840565b811461086057600080fd5b50565b6000813590506108728161084c565b92915050565b60006020828403121561088e5761088d610566565b5b600061089c84828501610863565b91505092915050565b6108ae81610590565b82525050565b60006040820190506108c960008301856108a5565b6108d66020830184610647565b9392505050565b6000815190506108ec816105db565b92915050565b60006020828403121561090857610907610566565b5b6000610916848285016108dd565b91505092915050565b600060208201905061093460008301846108a5565b92915050565b600081519050919050565b600082825260208201905092915050565b6000819050602082019050919050565b61096f816105ce565b82525050565b60006109818383610966565b60208301905092915050565b6000602082019050919050565b60006109a58261093a565b6109af8185610945565b93506109ba83610956565b8060005b838110156109eb5781516109d28882610975565b97506109dd8361098d565b9250506001810190506109be565b5085935050505092915050565b60006020820190508181036000830152610a12818461099a565b905092915050565b610a2381610840565b82525050565b6000602082019050610a3e6000830184610a1a565b9291505056fea2646970667358221220e87de3e2a5f13475da0ff801d19879653e5b48c9740a44125e08cd84270e866d64736f6c63430008120033
