--- hedera-node/test-clients/src/main/resources/contract/contracts/HRC904Reject/HRC904Reject.bin
@@ -1,1 +1,0 @@
-608060405234801561001057600080fd5b50600436106100365760003560e01c806376c6b3911461003b578063a869c78a14610059575b600080fd5b610043610089565b60405161005091906101a2565b60405180910390f35b610073600480360381019061006e9190610356565b610101565b60405161008091906101a2565b60405180910390f35b60003073ffffffffffffffffffffffffffffffffffffffff166376c6b3916040518163ffffffff1660e01b81526004016020604051808303816000875af11580156100d8573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906100fc91906103b4565b905090565b60003073ffffffffffffffffffffffffffffffffffffffff1663a869c78a836040518263ffffffff1660e01b815260040161013c919061049f565b6020604051808303816000875af115801561015b573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061017f91906103b4565b9050919050565b60008160070b9050919050565b61019c81610186565b82525050565b60006020820190506101b76000830184610193565b92915050565b6000604051905090565b600080fd5b600080fd5b600080fd5b6000601f19601f8301169050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b61021f826101d6565b810181811067ffffffffffffffff8211171561023e5761023d6101e7565b5b80604052505050565b60006102516101bd565b905061025d8282610216565b919050565b600067ffffffffffffffff82111561027d5761027c6101e7565b5b602082029050602081019050919050565b600080fd5b61029c81610186565b81146102a757600080fd5b50565b6000813590506102b981610293565b92915050565b60006102d26102cd84610262565b610247565b905080838252602082019050602084028301858111156102f5576102f461028e565b5b835b8181101561031e578061030a88826102aa565b8452602084019350506020810190506102f7565b5050509392505050565b600082601f83011261033d5761033c6101d1565b5b813561034d8482602086016102bf565b91505092915050565b60006020828403121561036c5761036b6101c7565b5b600082013567ffffffffffffffff81111561038a576103896101cc565b5b61039684828501610328565b91505092915050565b6000815190506103ae81610293565b92915050565b6000602082840312156103ca576103c96101c7565b5b60006103d88482850161039f565b91505092915050565b600081519050919050565b600082825260208201905092915050565b6000819050602082019050919050565b61041681610186565b82525050565b6000610428838361040d565b60208301905092915050565b6000602082019050919050565b600061044c826103e1565b61045681856103ec565b9350610461836103fd565b8060005b83811015610492578151610479888261041c565b975061048483610434565b925050600181019050610465565b5085935050505092915050565b600060208201905081810360008301526104b98184610441565b90509291505056fea264697066735822122090079297d3821301f7590ff582a2d38263cf93cb98de3e7f4d3246266bd4951364736f6c63430008120033
