--- hedera-node/test-clients/src/main/resources/contract/contracts/HRC904Reject/HRC904Reject.json
@@ -1,34 +1,0 @@
-[
-  {
-    "inputs": [],
-    "name": "rejectTokenFT",
-    "outputs": [
-      {
-        "internalType": "int64",
-        "name": "responseCode",
-        "type": "int64"
-      }
-    ],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "int64[]",
-        "name": "serialNumbers",
-        "type": "int64[]"
-      }
-    ],
-    "name": "rejectTokenNFTs",
-    "outputs": [
-      {
-        "internalType": "int64",
-        "name": "responseCode",
-        "type": "int64"
-      }
-    ],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  }
-]
