--- hedera-node/test-clients/src/main/resources/contract/contracts/HRC904TokenCancel/HRC904TokenCancel.bin
@@ -1,1 +1,0 @@
-608060405234801561001057600080fd5b506103a6806100206000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c8063ad4917cf1461003b578063cef5b7051461006b575b600080fd5b61005560048036038101906100509190610244565b61009b565b6040516100629190610293565b60405180910390f35b610085600480360381019061008091906102ae565b610123565b6040516100929190610293565b60405180910390f35b60003073ffffffffffffffffffffffffffffffffffffffff1663ad4917cf84846040518363ffffffff1660e01b81526004016100d89291906102ea565b6020604051808303816000875af11580156100f7573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061011b9190610328565b905092915050565b60003073ffffffffffffffffffffffffffffffffffffffff1663cef5b705836040518263ffffffff1660e01b815260040161015e9190610355565b6020604051808303816000875af115801561017d573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906101a19190610328565b9050919050565b600080fd5b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b60006101d8826101ad565b9050919050565b6101e8816101cd565b81146101f357600080fd5b50565b600081359050610205816101df565b92915050565b60008160070b9050919050565b6102218161020b565b811461022c57600080fd5b50565b60008135905061023e81610218565b92915050565b6000806040838503121561025b5761025a6101a8565b5b6000610269858286016101f6565b925050602061027a8582860161022f565b9150509250929050565b61028d8161020b565b82525050565b60006020820190506102a86000830184610284565b92915050565b6000602082840312156102c4576102c36101a8565b5b60006102d2848285016101f6565b91505092915050565b6102e4816101cd565b82525050565b60006040820190506102ff60008301856102db565b61030c6020830184610284565b9392505050565b60008151905061032281610218565b92915050565b60006020828403121561033e5761033d6101a8565b5b600061034c84828501610313565b91505092915050565b600060208201905061036a60008301846102db565b9291505056fea2646970667358221220f5eee7da8ac63e14ba98114de96d5ac10438085180aca5b0e83165db622198b264736f6c63430008120033
