--- hedera-node/test-clients/src/main/resources/contract/contracts/HRC904TokenClaim/HRC904TokenClaim.json
@@ -1,45 +1,0 @@
-[
-  {
-    "inputs": [
-      {
-        "internalType": "address",
-        "name": "sender",
-        "type": "address"
-      }
-    ],
-    "name": "claimAirdropFT",
-    "outputs": [
-      {
-        "internalType": "int64",
-        "name": "responseCode",
-        "type": "int64"
-      }
-    ],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "address",
-        "name": "sender",
-        "type": "address"
-      },
-      {
-        "internalType": "int64",
-        "name": "serial",
-        "type": "int64"
-      }
-    ],
-    "name": "claimAirdropNFT",
-    "outputs": [
-      {
-        "internalType": "int64",
-        "name": "responseCode",
-        "type": "int64"
-      }
-    ],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  }
-]
