--- hedera-node/test-clients/src/main/resources/contract/contracts/IHRC904UnlimitedAutoAssociations/IHRC904UnlimitedAutoAssociations.json
@@ -1,21 +1,0 @@
-[
-  {
-    "inputs": [
-      {
-        "internalType": "bool",
-        "name": "enableAutoAssociations",
-        "type": "bool"
-      }
-    ],
-    "name": "setUnlimitedAutomaticAssociations",
-    "outputs": [
-      {
-        "internalType": "int64",
-        "name": "responseCode",
-        "type": "int64"
-      }
-    ],
-    "stateMutability": "nonpayable",
-    "type": "function"
-  }
-]
