--- hedera-node/test-clients/src/main/resources/contract/contracts/IHRC904UnlimitedAutoAssociations/IHRC904UnlimitedAutoAssociations.sol
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-pragma solidity ^0.8.0;
-
-interface IHRC904UnlimitedAutoAssociations {
-    function setUnlimitedAutomaticAssociations(bool enableAutoAssociations) external returns (int64 responseCode);
-}
